requirements.txt
streamlit
numpy
pandas
onnxruntime
pillow
openpyxl